+++ src/plugin/src/reporting/metrics.js
@@ -1,5 +1,5 @@
-import { toArrayFromIterable } from "../array-utils.js";
-import { getNonEmptyString, normalizeStringList } from "../string-utils.js";
+import { toArrayFromIterable } from "../../../shared/array-utils.js";
+import { getNonEmptyString, normalizeStringList } from "../../../shared/string-utils.js";
 
 const hasHrtime = typeof process?.hrtime?.bigint === "function";
