--- src/shared/reporting.js
@@ -1,1 +1,0 @@
-export * from "./reporting/index.js";
